Carve out a better location for private driver backlight functionality
