+++ quantum/backlight/backlight_avr.c
@@ -1,10 +1,7 @@
 #include "quantum.h"
 #include "backlight.h"
+#include "backlight_driver_common.h"
 #include "debug.h"
-
-#if !defined(BACKLIGHT_PIN) && !defined(BACKLIGHT_PINS)
-#    error "Backlight pin/pins not defined. Please configure."
-#endif
 
 // This logic is a bit complex, we support 3 setups:
 //
