+++ quantum/backlight/backlight_driver_common.h
@@ -1,0 +1,7 @@
+#pragma once
+
+void backlight_pins_init(void);
+void backlight_pins_on(void);
+void backlight_pins_off(void);
+
+void breathing_task(void);
